Rename channel to client
A bit more clear since there are client and server-side classes
